requirements.txt
pytest
flake8
black
flake8-black
attrs